backend/requirements.txt
fastapi
uvicorn
python-multipart
librosa
soundfile
ffmpeg-python
requests
openai
yt-dlp
python-dotenv
razorpay
